refactor: moved generators from its package
