+++ src/core/generators/DoWhileTask.ts
@@ -1,4 +1,4 @@
-import { TaskType, DoWhileTaskDef } from "../types";
+import { TaskType, DoWhileTaskDef } from "../../common/types";
 import { DoWhileTaskDefGen, NestedTaskMapper } from "./types";
 import { nameTaskNameGenerator } from "./common";
