+++ src/core/generators/ForkJoinDynamicTask.ts
@@ -1,4 +1,4 @@
-import { ForkJoinDynamicDef, TaskType } from "../types";
+import { ForkJoinDynamicDef, TaskType } from "../../common/types";
 import { nameTaskNameGenerator } from "./common";
 
 export const generateForkJoinDynamic = (
