+++ src/core/generators/WaitTask.ts
@@ -1,4 +1,4 @@
-import { WaitTaskDef, TaskType } from "../types";
+import { WaitTaskDef, TaskType } from "../../common/types";
 import { nameTaskNameGenerator } from "./common";
 
 /**
