+++ src/core/generators/common.ts
